`plugins` is ambiguous. Changed to `exports`.
The `plugins` attribute hasn't caused any problems
so far, but it could in the future due to the package
itself being named `plugins`.

Right now, imports are used like: `from plugins import Plugin`

However, doing `import plugins` and `plugins.Plugin` could be a
problem if you are setting `plugins` to something else at the
end of the file.

For lack of a better name, `exports` is used instead.