Moved jquery download to a post plugin.
jQuery is downloaded to the correct directory,
but to ensure the directory exists or is created
we need to wait until create() is called.

Now that post plugins have access to the plugin
instance that created the content, the jquerydl
plugin will know when it is allowed to run (
only when the jquery plugin is used, and the
download is enabled).