Allow docopt for plugins args, remove ARGS hack.
If a plugin has a proper docopt usage string,
and cls.docopt == True, docopt will handle arg parsing
and the result is saved to self.argd. The original
arguments (a plugin's argv) is still saved under
self.args. If docopt isn't wanted (cls.docopt == False),
then self.argd is just a map of {flag_arg: True} for all
args that were passed (and begin with -).

Default args are used if no args were passed.

The ARGS[0] hack was removed. After some testing,
the circumstances that inspired that hack are not
producable with the new '--' separator for ARGS.
There is still some flip-flopping between PLUGIN and FILENAME,
because the current New usage string is ambiguos as far
as docopt is concerned.

It's easy to figure out if a FILENAME was meant to be a PLUGIN,
by trying to load it as a plugin.

If PLUGIN fails to load as a plugin name, then it is a FILENAME,
and loading by extension will work.

If both fail, then it wasn't a valid plugin name or file extension
to begin with.

If both PLUGIN and FILENAME are passed, there is no ambiguity.